Speeded up redrawing by a factor of bazillion
Apparently the correct way to redraw is to queue a redraw to gtk and
only draw during the expose event. Doing this sped up the drawing
considerably. (I only noticed this because I started wondering how can
resizing the window (which causes zooming) be so much smoother than
explicitly zooming...)